+++ anvil/modeling/docker/basepairmodels/basepairmodels/cli/shaputils.py
@@ -42,7 +42,8 @@ def combine_mult_and_diffref(mult, orig_inp, bg_data):
             
         to_return.append(np.mean(projected_hypothetical_contribs,axis=0))
         
-    to_return.append(np.zeros_like(orig_inp[1]))
+    if len(orig_inp) == 2:
+        to_return.append(np.zeros_like(orig_inp[1]))
     
     return to_return
 
@@ -176,14 +177,10 @@ def get_weightedsum_meannormed_logits(model, task_id, stranded):
     # See Google slide deck for explanations
     # We meannorm as per section titled 
     # "Adjustments for Softmax Layers" in the DeepLIFT paper
-    # Reshaping is done to be compatible with the single multinomial
-    reshaped_out=tf.reshape(model.outputs[0], [-1,model.outputs[0][:, :, start_idx:end_idx].shape[1]*(end_idx-start_idx)])
-
-    meannormed_logits=reshaped_out-tf.reduce_mean(reshaped_out, axis=1)[:,None]
-    # meannormed_logits = \
-    #     (model.outputs[0][:, :, start_idx:end_idx] - \
-    #      tf.reduce_mean(
-    #         model.outputs[0][:, :, start_idx:end_idx], axis=1)[:, None, :])
+    meannormed_logits = \
+        (model.outputs[0][:, :, start_idx:end_idx] - \
+         tf.reduce_mean(
+            model.outputs[0][:, :, start_idx:end_idx], axis=1)[:, None, :])
         
     # 'stop_gradient' will prevent importance from being propagated
     # through this operation; we do this because we just want to treat
@@ -197,11 +194,8 @@ def get_weightedsum_meannormed_logits(model, task_id, stranded):
     # Weight the logits according to the softmax probabilities, take
     # the sum for each example. This mirrors what was done for the
     # bpnet paper.
-    # weightedsum_meannormed_logits = tf.reduce_sum(softmax_out * \
-    #                                               meannormed_logits,
-    #                                               axis=(1, 2))
     weightedsum_meannormed_logits = tf.reduce_sum(softmax_out * \
                                                   meannormed_logits,
-                                                  axis=(1))
+                                                  axis=(1, 2))
     
     return weightedsum_meannormed_logits